+++ 09_Texturing/Sphere.cpp
@@ -5,19 +5,19 @@
 constexpr float M_PI = 3.14159265358979323846f;
 #endif
 
-Sphere::Sphere(const Vec3& center, double radius, const Material& material)
+Sphere::Sphere(const Vec3& center, float radius, const Material& material)
 	: Sphere(center, radius, material, Vec3{ 0, 0, 0 }, TextureCoordinates{ 1.0f, 1.0f }, TextureCoordinates{0.0f, 0.0f})
 {
 
 }
 
-Sphere::Sphere(const Vec3& center, double radius, const Material& material, const Vec3& rotation)
+Sphere::Sphere(const Vec3& center, float radius, const Material& material, const Vec3& rotation)
 	: Sphere(center, radius, material, rotation, TextureCoordinates{ 1.0f, 1.0f }, TextureCoordinates{ 0.0f, 0.0f })
 {
 
 }
 
-Sphere::Sphere(const Vec3& center, double radius, const Material& material, const Vec3& rotation, const TextureCoordinates& scale, const TextureCoordinates& bias)
+Sphere::Sphere(const Vec3& center, float radius, const Material& material, const Vec3& rotation, const TextureCoordinates& scale, const TextureCoordinates& bias)
 	: center(center), sqradius(radius*radius), material(material), rotation(rotation), scale(scale), bias(bias)
 {
 
@@ -32,16 +32,16 @@ std::optional<Intersection> Sphere::intersect(const Ray& ray) const
 {
 	Vec3 l = center - ray.getOrigin();
 
-	double tCenter = Vec3::dot(l, ray.getDirection());
+	float tCenter = Vec3::dot(l, ray.getDirection());
 	if (tCenter < 0)
 		return {};	// no intersection
 
-	double dSq = l.sqlength() - tCenter * tCenter;
+	float dSq = l.sqlength() - tCenter * tCenter;
 	if (dSq > sqradius)
 		return {};	// no intersection
 
-	double dist = sqrt(sqradius - dSq);
-	double t = tCenter - dist;
+	float dist = sqrt(sqradius - dSq);
+	float t = tCenter - dist;
 
 	if (t < 0)
 		t = tCenter + dist;	// when inside sphere
