+++ 09_Texturing/Sphere.h
@@ -5,16 +5,16 @@ class Sphere : public IntersectableObject
 {
 private:
 	const Vec3 center;
-	const double sqradius;
+	const float sqradius;
 	const Material material;
 	Vec3 rotation;
 	TextureCoordinates scale;
 	TextureCoordinates bias;
 
 public:
-	Sphere(const Vec3& center, double radius, const Material& material);
-	Sphere(const Vec3& center, double radius, const Material& material, const Vec3& rotation);
-	Sphere(const Vec3& center, double radius, const Material& material, const Vec3& rotation,
+	Sphere(const Vec3& center, float radius, const Material& material);
+	Sphere(const Vec3& center, float radius, const Material& material, const Vec3& rotation);
+	Sphere(const Vec3& center, float radius, const Material& material, const Vec3& rotation,
 		   const TextureCoordinates& scale, const TextureCoordinates& bias);
 
     virtual ~Sphere() {}
